LibGfx: Move VulkanImage into a separate file
