+++ Libraries/LibGfx/PaintingSurface.h
@@ -14,6 +14,14 @@
 #include <LibGfx/Size.h>
 #include <LibGfx/SkiaBackendContext.h>
 
+#ifdef USE_VULKAN_DMABUF_IMAGES
+namespace Gfx {
+
+struct VulkanImage;
+
+}
+#endif
+
 #ifdef AK_OS_MACOS
 #    include <LibGfx/MetalContext.h>
 #endif
